#2044: Moved CronAdjuster to package org.openhab.core.scheduler
