+++ bundles/org.openhab.core/src/main/java/org/openhab/core/scheduler/CronAdjuster.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.core.common;
+package org.openhab.core.scheduler;
 
 import java.time.DayOfWeek;
 import java.time.temporal.ChronoField;
@@ -30,7 +30,6 @@ import java.util.stream.IntStream;
 
 import org.eclipse.jdt.annotation.NonNullByDefault;
 import org.eclipse.jdt.annotation.Nullable;
-import org.openhab.core.scheduler.SchedulerTemporalAdjuster;
 
 /**
  * This class creates a {@link TemporalAdjuster} that takes a temporal and adjust it to the next deadline based on a
